fix(clients): sharpen migration decoder errors and docs accuracy
Make the generated JavaScript decoder's error prefix direction-neutral
(a future version is not stale), fail the hardening pass closed when any
migrationVersion decoder escapes wrapping, compare the CPI prefix oracle
offsets as usize, and correct the interactive migration page: keyboard
operable steps with aria-expanded state, 256 distinguishable versions,
and a --json example showing both streams. File #353 for the u64
configuration width that the runtime does not implement yet.

+++ crates/pina_cpi_renderer/src/__tests.rs
@@ -152,7 +152,11 @@ fn migration_version_is_part_of_the_framework_owned_cpi_prefix() {
 		else {
 			panic!("fixture discriminator is not a numeric constant");
 		};
-		assert_eq!(node.offset, bytes.len() as u64, "prefix must be contiguous");
+		assert_eq!(
+			node.offset as usize,
+			bytes.len(),
+			"prefix must be contiguous"
+		);
 		assert_eq!(r#type.format, U8, "fixture prefix is u8");
 		let Number::UnsignedInteger(byte) = value.number else {
 			panic!("fixture discriminator value is not an unsigned integer");
